Add FogOfWarController (recon)
This controller works by keeping a list of all unseen coordinates and
setting waypoints to one of them at a time, keeping track of which
coordinates have been seen along the way.

Slightly optimized the AStarController, which faster speed and
better logic.

Made additions to MetaController to help facilitate using various
controllers and to demo the FogOfWarController.