Add logic for getting objectives
i.e. keys and getting to the finish line.

Tuned the AStarController a bit more and fixed a couple of bugs in it.
Added a "hasArrived" method so outside classes can tell if the path is
complete. This has been added as part of the PathingStrategy interface.

Fixed a couple of bugs in the FogOfWarController.

Added another getEuclideanDistance method in utilities that takes 4
floats as input.